alpa_023_env_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=2.0.0=pypi_0
aiohttp=3.8.5=pypi_0
aiosignal=1.3.1=pypi_0
alpa=0.2.3=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2023.7.22=pypi_0
charset-normalizer=3.2.0=pypi_0
chex=0.1.5=pypi_0
click=8.0.4=pypi_0
contourpy=1.1.1=pypi_0
cupy-cuda113=10.6.0=pypi_0
cycler=0.11.0=pypi_0
datasets=2.14.5=pypi_0
dill=0.3.7=pypi_0
distlib=0.3.7=pypi_0
dm-tree=0.1.8=pypi_0
etils=1.3.0=pypi_0
fastrlock=0.8.2=pypi_0
filelock=3.12.4=pypi_0
flax=0.6.2=pypi_0
fonttools=4.42.1=pypi_0
frozenlist=1.4.0=pypi_0
fsspec=2023.6.0=pypi_0
grpcio=1.58.0=pypi_0
huggingface-hub=0.17.2=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.1.0=pypi_0
jax=0.3.22=pypi_0
jaxlib=0.3.22=dev_0
jsonschema=4.19.1=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
llvmlite=0.41.0=pypi_0
markdown-it-py=3.0.0=pypi_0
matplotlib=3.7.3=pypi_0
mdurl=0.1.2=pypi_0
ml-dtypes=0.2.0=pypi_0
msgpack=1.0.6=pypi_0
multidict=6.0.4=pypi_0
multiprocess=0.70.15=pypi_0
ncurses=6.3=h7f8727e_2
numba=0.58.0=pypi_0
numpy=1.23.0=pypi_0
openssl=1.1.1w=h7f8727e_0
opt-einsum=3.3.0=pypi_0
optax=0.1.7=pypi_0
packaging=23.1=pypi_0
pandas=2.0.3=pypi_0
pillow=10.0.1=pypi_0
pip=23.2.1=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=3.10.0=pypi_0
protobuf=4.24.3=pypi_0
pulp=2.7.0=pypi_0
pyarrow=13.0.0=pypi_0
pygments=2.16.1=pypi_0
pyparsing=3.1.1=pypi_0
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
ray=2.1.0=pypi_0
readline=8.1.2=h7f8727e_1
referencing=0.30.2=pypi_0
regex=2023.8.8=pypi_0
requests=2.31.0=pypi_0
rich=13.5.3=pypi_0
rpds-py=0.10.3=pypi_0
safetensors=0.3.3=pypi_0
scipy=1.10.1=pypi_0
setuptools=68.0.0=py38h06a4308_0
six=1.16.0=pypi_0
sqlite=3.38.5=hc218d9a_0
tensorstore=0.1.44=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
toolz=0.12.0=pypi_0
tqdm=4.66.1=pypi_0
transformers=4.33.2=pypi_0
typing-extensions=4.8.0=pypi_0
tzdata=2023.3=pypi_0
urllib3=2.0.5=pypi_0
virtualenv=20.24.5=pypi_0
wheel=0.38.4=py38h06a4308_0
xxhash=3.3.0=pypi_0
xz=5.2.5=h7f8727e_1
yarl=1.9.2=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.12=h7f8727e_2
